feat: 清除pyc文件

+++ clean_pyc.py
@@ -1,7 +1,10 @@
 import os
-path = __file__
+path = '.'
+env_dir = (path+os.path.sep+'.venv')
 for prefix, dirs, files in os.walk(path):
     for name in files:
         if name.endswith('.pyc'):
             filename = os.path.join(prefix, name)
+            if filename.startswith(env_dir):
+                continue
             os.remove(filename)
